fix: clean up docs react doctor findings
